Move pal and attestation types into separate directory
