+++ include/ccf/pal/attestation_sev_snp.h
