+++ include/ccf/pal/attestation_sgx.h
