+++ src/consensus/aft/impl/state.h
@@ -5,7 +5,7 @@
 #include "ccf/crypto/key_pair.h"
 #include "ccf/crypto/verifier.h"
 #include "ccf/ds/logger.h"
-#include "ccf/ds/pal.h"
+#include "ccf/pal.h"
 #include "ccf/tx_status.h"
 #include "consensus/aft/raft_types.h"
 #include "kv/kv_types.h"
