+++ src/ds/ring_buffer.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/ds/pal.h"
+#include "ccf/pal.h"
 #include "ring_buffer_types.h"
 
 #include <cstring>
