+++ src/node/jwt_key_auto_refresh.h
@@ -9,7 +9,7 @@
 #include "node/rpc/node_frontend.h"
 
 #define FMT_HEADER_ONLY
-#include "ccf/ds/pal.h"
+#include "ccf/pal.h"
 
 #include <fmt/format.h>
